% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomo-package.R
\docType{package}
\name{nomo-package}
\alias{nomo}
\alias{nomo-package}
\title{nomo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{sjobergd@mskcc.org}

Authors:
\itemize{
  \item Emily Vertosick \email{vertosie@mskcc.org}
}

}
\keyword{internal}
